Moved pgPool over to the models directory.
